+++ src/main/java/net/meander/subtlyd/mixin/common/world/entity/ai/goal/RandomStrollGoalMixin.java
@@ -1,5 +1,6 @@
 package net.meander.subtlyd.mixin.common.world.entity.ai.goal;
 
+import net.meander.subtlyd.world.entity.ai.goal.ShelteredRandomStrollGoal;
 import net.meander.subtlyd.world.entity.ai.util.GoalUtilsSD;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.Goal;
@@ -15,9 +16,10 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public abstract class RandomStrollGoalMixin extends Goal {
     @Shadow @Final protected PathfinderMob mob;
 
+    @SuppressWarnings("ConstantValue")
     @Inject(method = "canUse", at = @At("HEAD"), cancellable = true)
     private void maintainShelter(CallbackInfoReturnable<Boolean> cir) {
-        if (GoalUtilsSD.isMobSheltering(mob, mob.blockPosition())) {
+        if (!((Object) this instanceof ShelteredRandomStrollGoal) && GoalUtilsSD.isMobSheltering(mob, mob.blockPosition())) {
             cir.setReturnValue(false);
         }
     }