+++ src/main/java/net/meander/subtlyd/world/entity/LivingEntitySD.java
@@ -5,10 +5,7 @@ import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.Pose;
-import net.minecraft.world.entity.ai.goal.PanicGoal;
-import net.minecraft.world.entity.ai.goal.WrappedGoal;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.Vec3;
@@ -21,7 +18,6 @@ public abstract class LivingEntitySD extends LivingEntity {
     /**
      * Handles a tent sleep attempt. Searches for players that are sleeping within the specified tent.
      * @param tent Tent to test
-     * @param player Sleeping player
      */
     public static void startSleepingInTent(final Tent tent, final ServerPlayer player) {
         boolean foundSleepingPlayer = false;
@@ -52,20 +48,7 @@ public abstract class LivingEntitySD extends LivingEntity {
      * @param blockPos The position of the tent.
      * @param player The player to move.
      */
-    private static void setPosToTent(BlockPos blockPos, ServerPlayer player) {
+    default void setPosToTent(BlockPos blockPos, ServerPlayer player) {
         player.setPos(blockPos.getX() + 0.5, blockPos.getY(), blockPos.getZ() + 0.5);
-    }
-
-    /**
-     * @param mob The pathfinding mob to test.
-     * @return The speed multiplier for that animal type when it panics.
-     */
-    public static double getPanicSpeed(PathfinderMob mob) {
-        for (WrappedGoal wrappedGoal : mob.getGoalSelector().getAvailableGoals()) {
-            if (wrappedGoal.getGoal() instanceof PanicGoal panicGoal) {
-                return panicGoal.speedModifier;
-            }
-        }
-        return 1.25D;
     }
 }