+++ src/main/java/net/meander/subtlyd/world/entity/ai/goal/MoveToWarmthGoal.java
@@ -4,16 +4,12 @@ import net.meander.subtlyd.world.entity.ai.util.GoalUtilsSD;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.goal.MoveToBlockGoal;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelReader;
-import net.minecraft.world.level.LightLayer;
 
-public class SeekWarmthGoal extends MoveToBlockGoal {
-    protected final PathfinderMob mob;
-
-    public SeekWarmthGoal(PathfinderMob mob, double speedModifier) {
-        super(mob, speedModifier, 32, 16);
-
-        this.mob = mob;
+public class MoveToWarmthGoal extends MoveToBlockGoal {
+    public MoveToWarmthGoal(PathfinderMob mob, double speedModifier) {
+        super(mob, speedModifier, 16, 8);
     }
 
     @Override
@@ -23,11 +19,6 @@ public class SeekWarmthGoal extends MoveToBlockGoal {
 
     @Override
     protected boolean isValidTarget(LevelReader level, BlockPos pos) {
-        return level.getBrightness(LightLayer.BLOCK, pos) >= 10;
-    }
-
-    @Override
-    public double acceptedDistance() {
-        return 4.0;
+        return GoalUtilsSD.isWarmPos((Level) level, pos.above());
     }
 }