+++ src/main/java/net/meander/subtlyd/world/entity/ai/util/ConditionalPos.java
@@ -2,6 +2,7 @@ package net.meander.subtlyd.world.entity.ai.util;
 
 import net.minecraft.core.BlockPos;
 import net.minecraft.util.Mth;
+import net.minecraft.util.RandomSource;
 import net.minecraft.world.entity.PathfinderMob;
 import net.minecraft.world.entity.ai.util.DefaultRandomPos;
 import net.minecraft.world.phys.Vec3;
@@ -11,7 +12,10 @@ import java.util.function.BiPredicate;
 
 public class ConditionalPos {
     public static @Nullable BlockPos generateNearConditionalPos(final PathfinderMob mob, final int xzDist, final int yDist, final BiPredicate<PathfinderMob, BlockPos> predicate) {
-        for (int searchRadius = 8; searchRadius <= xzDist; searchRadius += 8) {
+        for (int searchRadius = 2; searchRadius <= xzDist; searchRadius += 2) {
+            RandomSource random = mob.getRandom();
+            searchRadius = random.nextBoolean() ? searchRadius : searchRadius + 2;
+
             for (int attempts = 0; attempts < searchRadius * Mth.sqrt(searchRadius); attempts++) {
                 Vec3 randPos = DefaultRandomPos.getPos(mob, searchRadius, yDist);