+++ docs/RELEASE_NOTES.md
@@ -52,6 +52,7 @@ Developer Note: Did you know this mod has over 200 Java files! I sure didn't unt
 - Added `arrow_flammable` block tag
 - Added `causes_flock_panic` damage type tag
 - Changed `can_be_full` to be named `predator`
+- Changed `can_be_scared` to `panics_with_flock`
 - Changed `can_seek_shade` to `seeks_shade`
 - Changed `can_seek_warmth` to `seeks_warmth`
 - Removed `can_break_tents` damage type tag